- Adding Streaming on dashboard page: There are 2 ways to use streaming, the first is using the `loading.tsx` in a page level, the second is granularly adding a `<Suspense />` on the Page components.
- Adding a group folder for Dashboard: when we create a `loading.tsx`it will allow the how Dashboard page to render a "loading" ui for the whole page, but it can lead to a bug for pages under the "Dashboard" page that the loading would be shown even on them, therefore a new folder, under app/dashboard/ was create to be a "RouteGroup", so only dashboard/(overview)/page  is shown with the "loading".

- Added every Dashboard component, a `<Suspense />`: it will allow granulated streaming, giving to the end user a delightful loading experience for every major component.
